chore: allow measuring performance of SBOM ingestion
