+++ modules/graph/Cargo.toml
@@ -10,6 +10,7 @@ trustify-common = { workspace = true }
 trustify-migration = { workspace = true }
 trustify-cvss = { workspace = true }
 trustify-module-search = { workspace = true }
+
 actix-web = { workspace = true }
 anyhow = { workspace = true }
 async-trait = { workspace = true }
@@ -31,7 +32,10 @@ tempfile = { workspace = true }
 thiserror = { workspace = true }
 time = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
+tracing = { workspace = true }
 utoipa = { workspace = true, features = ["actix_extras"] }
 
 [dev-dependencies]
+humantime = "2"
+rust-lzma = "0.6"
 test-log = { workspace = true, features = ["env_logger", "trace"] }
