+++ modules/graph/src/graph/sbom/mod.rs
@@ -14,6 +14,7 @@ use sea_query::{Condition, Func, JoinType, Query, SimpleExpr};
 use std::collections::{HashMap, HashSet};
 use std::fmt::{Debug, Formatter};
 use time::OffsetDateTime;
+use tracing::instrument;
 use trustify_common::cpe::Cpe;
 use trustify_common::db::Transactional;
 use trustify_common::package::PackageVulnerabilityAssertions;
@@ -23,6 +24,7 @@ use trustify_entity as entity;
 use trustify_entity::relationship::Relationship;
 
 pub mod spdx;
+mod tests;
 
 #[derive(Clone, Default)]
 pub struct SbomInformation {
@@ -50,19 +52,22 @@ impl Graph {
             .map(|sbom| (self, sbom).into()))
     }
 
-    pub async fn get_sbom(
+    #[instrument(skip(tx))]
+    pub async fn get_sbom<TX: AsRef<Transactional>>(
         &self,
         location: &str,
         sha256: &str,
+        tx: TX,
     ) -> Result<Option<SbomContext>, Error> {
         Ok(entity::sbom::Entity::find()
             .filter(Condition::all().add(entity::sbom::Column::Location.eq(location)))
             .filter(Condition::all().add(entity::sbom::Column::Sha256.eq(sha256.to_string())))
-            .one(&self.db)
+            .one(&self.connection(&tx))
             .await?
             .map(|sbom| (self, sbom).into()))
     }
 
+    #[instrument(skip(tx, info), err)]
     pub async fn ingest_sbom<TX: AsRef<Transactional>>(
         &self,
         location: &str,
@@ -71,7 +76,7 @@ impl Graph {
         info: impl Into<SbomInformation>,
         tx: TX,
     ) -> Result<SbomContext, Error> {
-        if let Some(found) = self.get_sbom(location, sha256).await? {
+        if let Some(found) = self.get_sbom(location, sha256, &tx).await? {
             return Ok(found);
         }
 
@@ -88,7 +93,7 @@ impl Graph {
             ..Default::default()
         };
 
-        Ok((self, model.insert(&self.db).await?).into())
+        Ok((self, model.insert(&self.connection(&tx)).await?).into())
     }
 
     /// Fetch a single SBOM located via internal `id`, external `location` (URL),
@@ -345,7 +350,8 @@ impl From<(&Graph, entity::sbom::Model)> for SbomContext {
 }
 
 impl SbomContext {
-    pub async fn ingest_describes_cpe22<C: Into<Cpe>, TX: AsRef<Transactional>>(
+    #[instrument(skip(tx), err)]
+    pub async fn ingest_describes_cpe22<C: Into<Cpe> + Debug, TX: AsRef<Transactional>>(
         &self,
         cpe: C,
         tx: TX,
@@ -369,6 +375,7 @@ impl SbomContext {
         Ok(())
     }
 
+    #[instrument(skip(tx), err)]
     pub async fn ingest_describes_package<TX: AsRef<Transactional>>(
         &self,
         purl: Purl,
@@ -395,6 +402,7 @@ impl SbomContext {
         Ok(())
     }
 
+    #[instrument(skip(tx), err)]
     pub async fn describes_packages<TX: AsRef<Transactional>>(
         &self,
         tx: TX,
@@ -411,6 +419,7 @@ impl SbomContext {
             .await
     }
 
+    #[instrument(skip(tx), err)]
     pub async fn describes_cpe22s<TX: AsRef<Transactional>>(
         &self,
         tx: TX,
@@ -429,6 +438,7 @@ impl SbomContext {
 
     /// Within the context of *this* SBOM, ingest a relationship between
     /// two packages.
+    #[instrument(skip(tx), err)]
     async fn ingest_package_relates_to_package<TX: AsRef<Transactional>>(
         &self,
         left_package_input: Purl,
@@ -688,548 +698,6 @@ impl SbomContext {
             .await?;
 
         Ok(packages_to_purls(found)?)
-    }
-
-     */
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::graph::Graph;
-    use test_log::test;
-    use trustify_common::db::{Database, Transactional};
-    use trustify_common::purl::Purl;
-    use trustify_common::sbom::SbomLocator;
-    use trustify_entity::relationship::Relationship;
-
-    #[test(tokio::test)]
-    async fn ingest_sboms() -> Result<(), anyhow::Error> {
-        let db = Database::for_test("ingest_sboms").await?;
-        let system = Graph::new(db);
-
-        let sbom_v1 = system
-            .ingest_sbom(
-                "http://sbom.com/test.json",
-                "8",
-                "a",
-                (),
-                Transactional::None,
-            )
-            .await?;
-        let sbom_v1_again = system
-            .ingest_sbom(
-                "http://sbom.com/test.json",
-                "8",
-                "b",
-                (),
-                Transactional::None,
-            )
-            .await?;
-        let sbom_v2 = system
-            .ingest_sbom(
-                "http://sbom.com/test.json",
-                "9",
-                "c",
-                (),
-                Transactional::None,
-            )
-            .await?;
-
-        let other_sbom = system
-            .ingest_sbom(
-                "http://sbom.com/other.json",
-                "10",
-                "d",
-                (),
-                Transactional::None,
-            )
-            .await?;
-
-        assert_eq!(sbom_v1.sbom.id, sbom_v1_again.sbom.id);
-
-        assert_ne!(sbom_v1.sbom.id, sbom_v2.sbom.id);
-        Ok(())
-    }
-
-    #[test(tokio::test)]
-    async fn ingest_and_fetch_sboms_describing_purls() -> Result<(), anyhow::Error> {
-        let db = Database::for_test("ingest_and_fetch_sboms_describing_purls").await?;
-        let system = Graph::new(db);
-
-        let sbom_v1 = system
-            .ingest_sbom(
-                "http://sbom.com/test.json",
-                "8",
-                "a",
-                (),
-                Transactional::None,
-            )
-            .await?;
-        let sbom_v2 = system
-            .ingest_sbom(
-                "http://sbom.com/test.json",
-                "9",
-                "b",
-                (),
-                Transactional::None,
-            )
-            .await?;
-        let sbom_v3 = system
-            .ingest_sbom(
-                "http://sbom.com/test.json",
-                "10",
-                "c",
-                (),
-                Transactional::None,
-            )
-            .await?;
-
-        sbom_v1
-            .ingest_describes_package(
-                "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        sbom_v2
-            .ingest_describes_package(
-                "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        sbom_v3
-            .ingest_describes_package(
-                "pkg://maven/io.quarkus/quarkus-core@1.9.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        let found = system
-            .locate_sboms(
-                SbomLocator::Purl("pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?),
-                Transactional::None,
-            )
-            .await?;
-
-        assert_eq!(2, found.len());
-        assert!(found.contains(&sbom_v1));
-        assert!(found.contains(&sbom_v2));
-
-        Ok(())
-    }
-
-    #[test(tokio::test)]
-    async fn ingest_and_locate_sboms_describing_cpes() -> Result<(), anyhow::Error> {
-        let db = Database::for_test("ingest_and_locate_sboms_describing_cpes").await?;
-        let system = Graph::new(db);
-
-        let sbom_v1 = system
-            .ingest_sbom(
-                "http://sbom.com/test.json",
-                "8",
-                "a",
-                (),
-                Transactional::None,
-            )
-            .await?;
-        let sbom_v2 = system
-            .ingest_sbom(
-                "http://sbom.com/test.json",
-                "9",
-                "b",
-                (),
-                Transactional::None,
-            )
-            .await?;
-        let sbom_v3 = system
-            .ingest_sbom(
-                "http://sbom.com/test.json",
-                "10",
-                "c",
-                (),
-                Transactional::None,
-            )
-            .await?;
-
-        sbom_v1
-            .ingest_describes_cpe22(
-                cpe::uri::Uri::parse("cpe:/a:redhat:quarkus:2.13::el8")?,
-                Transactional::None,
-            )
-            .await?;
-
-        sbom_v2
-            .ingest_describes_cpe22(
-                cpe::uri::Uri::parse("cpe:/a:redhat:quarkus:2.13::el8")?,
-                Transactional::None,
-            )
-            .await?;
-
-        sbom_v3
-            .ingest_describes_cpe22(
-                cpe::uri::Uri::parse("cpe:/a:redhat:not-quarkus:2.13::el8")?,
-                Transactional::None,
-            )
-            .await?;
-
-        let found = system
-            .locate_sboms(
-                SbomLocator::Cpe(cpe::uri::Uri::parse("cpe:/a:redhat:quarkus:2.13::el8")?.into()),
-                Transactional::None,
-            )
-            .await?;
-
-        assert_eq!(2, found.len());
-        assert!(found.contains(&sbom_v1));
-        assert!(found.contains(&sbom_v2));
-
-        Ok(())
-    }
-
-    #[test(tokio::test)]
-    async fn transitive_dependency_of() -> Result<(), anyhow::Error> {
-        let db = Database::for_test("transitive_dependency_of").await?;
-        let system = Graph::new(db);
-
-        let sbom1 = system
-            .ingest_sbom(
-                "http://sbomsRus.gov/thing1.json",
-                "8675309",
-                "a",
-                (),
-                Transactional::None,
-            )
-            .await?;
-
-        sbom1
-            .ingest_package_relates_to_package(
-                "pkg://maven/io.quarkus/transitive-b@1.2.3".try_into()?,
-                Relationship::DependencyOf,
-                "pkg://maven/io.quarkus/transitive-a@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        sbom1
-            .ingest_package_relates_to_package(
-                "pkg://maven/io.quarkus/transitive-c@1.2.3".try_into()?,
-                Relationship::DependencyOf,
-                "pkg://maven/io.quarkus/transitive-b@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        sbom1
-            .ingest_package_relates_to_package(
-                "pkg://maven/io.quarkus/transitive-d@1.2.3".try_into()?,
-                Relationship::DependencyOf,
-                "pkg://maven/io.quarkus/transitive-c@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        sbom1
-            .ingest_package_relates_to_package(
-                "pkg://maven/io.quarkus/transitive-e@1.2.3".try_into()?,
-                Relationship::DependencyOf,
-                "pkg://maven/io.quarkus/transitive-c@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        sbom1
-            .ingest_package_relates_to_package(
-                "pkg://maven/io.quarkus/transitive-d@1.2.3".try_into()?,
-                Relationship::DependencyOf,
-                "pkg://maven/io.quarkus/transitive-b@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        let results = sbom1
-            .related_packages_transitively(
-                &[Relationship::DependencyOf],
-                "pkg://maven/io.quarkus/transitive-a@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        Ok(())
-    }
-
-    #[test(tokio::test)]
-    async fn ingest_package_relates_to_package_dependency_of() -> Result<(), anyhow::Error> {
-        let db = Database::for_test("ingest_contains_packages").await?;
-        let system = Graph::new(db);
-
-        let sbom1 = system
-            .ingest_sbom(
-                "http://sbomsRus.gov/thing1.json",
-                "8675309",
-                "a",
-                (),
-                Transactional::None,
-            )
-            .await?;
-
-        sbom1
-            .ingest_package_relates_to_package(
-                "pkg://maven/io.quarkus/quarkus-postgres@1.2.3".try_into()?,
-                Relationship::DependencyOf,
-                "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        let sbom2 = system
-            .ingest_sbom(
-                "http://sbomsRus.gov/thing2.json",
-                "8675308",
-                "b",
-                (),
-                Transactional::None,
-            )
-            .await?;
-
-        sbom2
-            .ingest_package_relates_to_package(
-                "pkg://maven/io.quarkus/quarkus-sqlite@1.2.3".try_into()?,
-                Relationship::DependencyOf,
-                "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        let dependencies = sbom1
-            .related_packages(
-                Relationship::DependencyOf,
-                "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        assert_eq!(1, dependencies.len());
-
-        assert_eq!(
-            "pkg://maven/io.quarkus/quarkus-postgres@1.2.3",
-            Purl::from(dependencies[0].clone()).to_string()
-        );
-
-        let dependencies = sbom2
-            .related_packages(
-                Relationship::DependencyOf,
-                "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-                Transactional::None,
-            )
-            .await?;
-
-        assert_eq!(1, dependencies.len());
-
-        assert_eq!(
-            "pkg://maven/io.quarkus/quarkus-sqlite@1.2.3",
-            Purl::from(dependencies[0].clone()).to_string()
-        );
-
-        Ok(())
-    }
-
-    #[test(tokio::test)]
-    async fn sbom_vulnerabilities() -> Result<(), anyhow::Error> {
-        let db = Database::for_test("sbom_vulnerabilities").await?;
-        let system = Graph::new(db);
-
-        println!("{:?}", system);
-
-        let sbom = system
-            .ingest_sbom(
-                "http://sbomsRus.gov/thing1.json",
-                "8675309",
-                "a",
-                (),
-                Transactional::None,
-            )
-            .await?;
-
-        println!("-------------------- A");
-
-        sbom.ingest_describes_package("pkg://oci/my-app@1.2.3".try_into()?, Transactional::None)
-            .await?;
-        println!("-------------------- B");
-
-        sbom.ingest_package_relates_to_package(
-            "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-            Relationship::DependencyOf,
-            "pkg://oci/my-app@1.2.3".try_into()?,
-            Transactional::None,
-        )
-        .await?;
-        println!("-------------------- C");
-
-        sbom.ingest_package_relates_to_package(
-            "pkg://maven/io.quarkus/quarkus-postgres@1.2.3".try_into()?,
-            Relationship::DependencyOf,
-            "pkg://maven/io.quarkus/quarkus-core@1.2.3".try_into()?,
-            Transactional::None,
-        )
-        .await?;
-        println!("-------------------- D");
-
-        sbom.ingest_package_relates_to_package(
-            "pkg://maven/postgres/postgres-driver@1.2.3".try_into()?,
-            Relationship::DependencyOf,
-            "pkg://maven/io.quarkus/quarkus-postgres@1.2.3".try_into()?,
-            Transactional::None,
-        )
-        .await?;
-
-        let advisory = system
-            .ingest_advisory(
-                "RHSA-1",
-                "http://redhat.com/secdata/RHSA-1",
-                "7",
-                (),
-                Transactional::None,
-            )
-            .await?;
-
-        let advisory_vulnerability = advisory
-            .link_to_vulnerability("CVE-00000001", Transactional::None)
-            .await?;
-
-        advisory_vulnerability
-            .ingest_affected_package_range(
-                "pkg://maven/postgres/postgres-driver".try_into()?,
-                "1.1",
-                "1.9",
-                Transactional::None,
-            )
-            .await?;
-
-        let assertions = sbom.vulnerability_assertions(Transactional::None).await?;
-
-        assert_eq!(1, assertions.len());
-
-        let affected_purls = assertions
-            .keys()
-            .map(|e| Purl::from(e.clone()))
-            .collect::<Vec<_>>();
-
-        assert_eq!(
-            affected_purls[0].to_string(),
-            "pkg://maven/postgres/postgres-driver@1.2.3"
-        );
-
-        Ok(())
-    }
-
-    /*
-    #[tokio::test]
-    async fn ingest_contains_packages() -> Result<(), anyhow::Error> {
-        env_logger::builder()
-        .filter_level(log::LevelFilter::Info)
-        .is_test(true)
-        .init();
-
-            let graph = InnerSystem::for_test("ingest_contains_packages").await?;
-
-            let sbom = graph
-                .ingest_sbom("http://sboms.mobi/something.json", "7", Transactional::None)
-                .await?;
-
-            let contains1 = sbom
-                .ingest_contains_package(
-                    "pkg://maven/io.quarkus/quarkus-core@1.2.3",
-                    Transactional::None,
-                )
-                .await?;
-
-            let contains2 = sbom
-                .ingest_contains_package(
-                    "pkg://maven/io.quarkus/quarkus-core@1.2.3",
-                    Transactional::None,
-                )
-                .await?;
-
-            let contains3 = sbom
-                .ingest_contains_package(
-                    "pkg://maven/io.quarkus/quarkus-addons@1.2.3",
-                    Transactional::None,
-                )
-                .await?;
-
-            assert_eq!(
-                contains1.sbom_contains_package.qualified_package_id,
-                contains2.sbom_contains_package.qualified_package_id
-            );
-            assert_ne!(
-                contains1.sbom_contains_package.qualified_package_id,
-                contains3.sbom_contains_package.qualified_package_id
-            );
-
-            let mut contains = sbom.contains_packages(Transactional::None).await?;
-
-            assert_eq!(2, contains.len());
-
-            let contains: Vec<_> = contains.drain(0..).map(Purl::from).collect();
-
-            assert!(contains.contains(&Purl::from("pkg://maven/io.quarkus/quarkus-core@1.2.3")));
-            assert!(contains.contains(&Purl::from("pkg://maven/io.quarkus/quarkus-addons@1.2.3")));
-
-            Ok(())
-        }
-         */
-
-    /*
-
-    #[tokio::test]
-    async fn ingest_and_fetch_sbom_packages() -> Result<(), anyhow::Error> {
-        /*
-        env_logger::builder()
-            .filter_level(log::LevelFilter::Info)
-            .is_test(true)
-            .init();
-
-         */
-        let graph = InnerSystem::for_test("ingest_and_fetch_sbom_packages").await?;
-
-        let sbom_v1 = graph.ingest_sbom("http://sbom.com/test.json", "8").await?;
-        let sbom_v2 = graph.ingest_sbom("http://sbom.com/test.json", "9").await?;
-        let sbom_v3 = graph
-            .ingest_sbom("http://sbom.com/test.json", "10")
-            .await?;
-
-        sbom_v1
-            .ingest_sbom_dependency("pkg://maven/io.quarkus/taco@1.2.3", Transactional::None)
-            .await?;
-
-        sbom_v1
-            .ingest_package_dependency(
-                "pkg://maven/io.quarkus/foo@1.2.3",
-                "pkg://maven/io.quarkus/baz@1.2.3",
-                Transactional::None,
-            )
-            .await?;
-
-        sbom_v2
-            .ingest_package_dependency(
-                "pkg://maven/io.quarkus/foo@1.2.3",
-                "pkg://maven/io.quarkus/bar@1.2.3",
-                Transactional::None,
-            )
-            .await?;
-
-        let sbom_packages = sbom_v1.all_packages(Transactional::None).await?;
-        assert_eq!(3, sbom_packages.len());
-
-        for sbom_package in sbom_packages {
-            let _sboms = sbom_package
-                .package
-                .sboms_containing(Transactional::None)
-                .await?;
-        }
-
-        Ok(())
     }
 
      */
